backend/requirements.txt
flask>=2.0.1
flask-sqlalchemy>=2.5.1
flask-cors>=3.0.10
Pillow>=8.3.1
werkzeug>=2.0.1 